# backend/swagger/docs/swagger.yaml
definitions:
...
  users.AuthToken:
    properties:
      Permissions:
        $ref: '#/definitions/users.Permissions'
      belongsTo:
        description: legacy stored metadata only, never trusted for auth; identity
          is always the hashed_tokens lookup
        type: integer
      expiresAt:
        type: integer
      issuedAt:
        type: integer
      key:
        description: for backward compatibility
        type: string
      name:
        type: string
      token:
        type: string
      username:
        type: string
    type: object
  users.BackendScope:
    properties:
      path:
        description: real path for the source
        type: string
      permissions:
        $ref: '#/definitions/users.SourceFilePermissions'
      scope:
        description: index path within that source
        type: string
    type: object
  users.FileLoading:
    properties:
      clearAll:
        type: boolean
      downloadChunkSizeMb:
        type: integer
      maxConcurrentUpload:
        type: integer
      uploadChunkSizeMb:
        type: integer
    type: object
  users.FrontendScope:
    properties:
      name:
        description: 'Bolt: filesystem path; JSON API: display name after prepForFrontend'
        type: string
      permissions:
        $ref: '#/definitions/users.SourceFilePermissions'
      scope:
        description: index path within that source
        type: string
    type: object
  users.LoginMethod:
    enum:
    - password
    - proxy
    - oidc
    - ldap
    - jwt
    type: string
    x-enum-varnames:
    - LoginMethodPassword
    - LoginMethodProxy
    - LoginMethodOidc
    - LoginMethodLdap
    - LoginMethodJwt
  users.Permissions:
    properties:
      admin:
        description: allow admin access
        type: boolean
      api:
        description: allow api access
        type: boolean
      create:
        description: 'deprecated: legacy user migration only; not used for API tokens'
        type: boolean
      delete:
        description: 'deprecated: legacy user migration only; not used for API tokens'
        type: boolean
      download:
        description: 'deprecated: legacy user migration only; not used for API tokens'
        type: boolean
      modify:
        description: 'deprecated: legacy user migration only; not used for API tokens'
        type: boolean
      realtime:
        description: allow realtime updates
        type: boolean
      share:
        description: allow sharing files
        type: boolean
      view:
        description: 'deprecated: legacy user migration only; not used for API tokens'
        type: boolean
    type: object
  users.PinnedItems:
    additionalProperties:
      additionalProperties:
        items:
          type: string
        type: array
      type: object
    type: object
...
  users.SourceFilePermissions:
    properties:
      configured:
        description: true when explicitly set (allows intentional deny-all)
        type: boolean
      create:
        type: boolean
      delete:
        type: boolean
      download:
        type: boolean
      modify:
        type: boolean
      view:
        type: boolean
    type: object
  users.User:
    properties:
      apiKeys:
        additionalProperties:
          $ref: '#/definitions/users.AuthToken'
        type: object
      avatarUrl:
        type: string
      backendScopes:
        description: |-
          BackendScopes is the authoritative, persisted access list (SourceScope.Name = backend source path).
          SQLite stores this inside user_data JSON under the key "scopes" (see sqldb.UserData).
        items:
          $ref: '#/definitions/users.BackendScope'
        type: array
      backendSourcePermissions:
        additionalProperties:
          $ref: '#/definitions/users.SourceFilePermissions'
        description: key = source path
        type: object
      customTheme:
        description: Name of theme to use chosen from custom themes config.
        type: string
      darkMode:
        description: should dark mode be enabled
        type: boolean
      dateFormat:
        description: when false, the date is relative, when true, the date is an exact
          timestamp
        type: boolean
      debugOffice:
        description: debug onlyoffice editor
        type: boolean
      deleteAfterArchive:
        description: delete source files after successful creation/extraction of archives
        type: boolean
      deleteWithoutConfirming:
        description: delete files without confirmation
        type: boolean
      disableOfficePreviewExt:
        description: deprecated
        type: string
      disableOnlyOfficeExt:
        description: deprecated
        type: string
      disablePreviewExt:
        description: space separated list of file extensions to disable preview for
        type: string
      disableQuickToggles:
        description: disable the quick toggles in the sidebar
        type: boolean
      disableSearchOptions:
        description: disable the search options in the search bar
        type: boolean
      disableSettings:
        type: boolean
      disableUpdateNotifications:
        description: disable update notifications
        type: boolean
      disableViewingExt:
        description: space separated list of file extensions to disable viewing for
        type: string
      editorQuickSave:
        description: show quick save button in editor
        type: boolean
      fileLoading:
        allOf:
        - $ref: '#/definitions/users.FileLoading'
        description: upload and download settings
      gallerySize:
        description: 0-9 - the size of the gallery thumbnails
        type: integer
      hideFileExt:
        description: space separated list of file extensions to hide in UI and API
        type: string
      hideFilesInTree:
        description: hide files in the sidebar tree navigation, when true, will show
          only directories.
        type: boolean
      hideSidebarFileActions:
        description: hide the file actions in the sidebar
        type: boolean
      id:
        type: integer
      locale:
        description: 'language to use: eg. de, en, or fr'
        type: string
      lockPassword:
        type: boolean
      loginMethod:
        $ref: '#/definitions/users.LoginMethod'
      otpEnabled:
        type: boolean
      passkeyCredentials:
        items:
          $ref: '#/definitions/users.WebAuthnCredential'
        type: array
      password:
        type: string
      perm:
        $ref: '#/definitions/users.Permissions'
      permissions:
        allOf:
        - $ref: '#/definitions/users.Permissions'
        description: 'global: admin, api, share, realtime'
      pinnedItems:
        $ref: '#/definitions/users.PinnedItems'
      preferEditorForMarkdown:
        description: prefer editor first for markdown files instead of the Markdown
          Viewer
        type: boolean
      preview:
        $ref: '#/definitions/users.Preview'
      quickDownload:
        description: show icon to download in one click
        type: boolean
      scopes:
        items:
          $ref: '#/definitions/users.FrontendScope'
        type: array
      showCopyPath:
        description: show copy path action in the context menu
        type: boolean
      showFirstLogin:
        type: boolean
      showHidden:
        description: show hidden files in the UI. On windows this includes files starting
          with a dot and windows hidden files
        type: boolean
      showSelectMultiple:
        description: show select multiple files on desktop
        type: boolean
      showToolsInSidebar:
        description: 'when false, sidebar hides links with category "tool" (default:
          true)'
        type: boolean
      sidebarLinks:
        description: customizable sidebar links
        items:
          $ref: '#/definitions/users.SidebarLink'
        type: array
      singleClick:
        description: open directory on single click, also enables middle click to
          open in new tab
        type: boolean
      sorting:
        $ref: '#/definitions/users.Sorting'
      sourcePermissions:
        additionalProperties:
          $ref: '#/definitions/users.SourceFilePermissions'
        description: 'deprecated: use scopes[].permissions'
        type: object
      stickySidebar:
        description: keep sidebar open when navigating
        type: boolean
      themeColor:
        description: 'theme color to use: eg. #ff0000, or var(--red), var(--purple),
          etc'
        type: string
      tokens:
        additionalProperties:
          $ref: '#/definitions/users.AuthToken'
        type: object
      totpNonce:
        type: string
      totpSecret:
        type: string
      username:
        type: string
      version:
        type: integer
      viewMode:
        description: 'view mode to use: eg. normal, list, grid, or compact'
        type: string
    type: object
  users.WebAuthnCredential:
    properties:
      aaguid:
        type: string
      attestationFormat:
        type: string
      attestationObj:
        type: string
      attestationType:
        type: string
      authenticatorData:
        type: string
      clientDataHash:
        type: string
      clientDataJSON:
        type: string
      cloneWarning:
        type: boolean
      createdAt:
        type: integer
      flags:
        $ref: '#/definitions/users.WebAuthnCredentialFlags'
      id:
        type: string
      lastUsedAt:
        type: integer
      name:
        type: string
      publicKey:
        type: string
      publicKeyAlg:
        type: integer
      signCount:
        type: integer
      transport:
        items:
          type: string
        type: array
    type: object
...
  web.UserSearchResult:
    properties:
      avatarUrl:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/users:
    delete:
      consumes:
      - application/json
      description: Deletes a user identified by login name.
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user by username
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Returns users or only the current user; with ?q=<text>, users with
        share permission receive only matching usernames and avatar URLs for share
        selectors. With ?username=self, the logged-in user; with ?username=login,
        that user if permitted. Query id= is not supported.
      parameters:
      - description: Login name, or 'self' for the current session user
        in: query
        name: username
        type: string
      - description: Username substring for the share recipient selector (requires
          share permission)
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details, matching selector results, or user list
          schema:
            $ref: '#/definitions/web.UserSearchResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List users or get one by username
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates the details of a user identified by ID. which must list
        the JSON field names to update. When the authenticated actor uses password
        login, they must send their current password in the X-Password header unless
        the update only touches NonAdminEditable profile fields (not password). Any
        admin-only field requires confirmation.
      parameters:
      - description: user ID to update
        in: query
        name: id
        type: string
      - description: usename to update
        in: query
        name: id
        type: string
      - description: Actor's current password (URL-encoded); required for password-login
          actors when updating password or any field outside NonAdminEditable
        in: header
        name: X-Password
        type: string
      - description: User data to update
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/users.User'
      responses:
        "204":
          description: No Content - User updated successfully
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - invalid or missing actor password when required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user's details
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Adds a new user to the system. When the authenticated actor uses
        password login, they must send their current password in the X-Password header.
      parameters:
      - description: Actor's current password (URL-encoded); required for password-login
          actors
        in: header
        name: X-Password
        type: string
      - description: User data to create a new user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/users.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created user
          schema:
            $ref: '#/definitions/users.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - invalid or missing actor password when required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - Users
...
swagger: "2.0"